make a deserialize factory
