+++ tests/Speckle.Sdk.Serialization.Testing/Program.cs
@@ -1,16 +1,13 @@
 #pragma warning disable CA1506
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging.Abstractions;
 using Speckle.Sdk;
 using Speckle.Sdk.Credentials;
 using Speckle.Sdk.Host;
 using Speckle.Sdk.Models;
 using Speckle.Sdk.Serialisation.V2;
-using Speckle.Sdk.Serialisation.V2.Receive;
 using Speckle.Sdk.Serialisation.V2.Send;
 using Speckle.Sdk.Serialization.Testing;
-using Speckle.Sdk.SQLite;
 
 const bool skipCacheReceive = false;
 const bool skipCacheSendCheck = true;
@@ -41,21 +38,15 @@ Console.WriteLine("Attach");
 var token = serviceProvider.GetRequiredService<IAccountManager>().GetDefaultAccount()?.token;
 var progress = new Progress(true);
 
-var factory = new SerializeProcessFactory(
-  new BaseChildFinder(new BasePropertyGatherer()),
-  new ObjectSerializerFactory(new BasePropertyGatherer()),
-  new BaseDeserializer(new ObjectDeserializerFactory()),
-  serviceProvider.GetRequiredService<ISqLiteJsonCacheManagerFactory>(),
-  serviceProvider.GetRequiredService<IServerObjectManagerFactory>(),
-  new NullLoggerFactory()
-);
+var factory = serviceProvider.GetRequiredService<IDeserializeProcessFactory>();
 var process = factory.CreateDeserializeProcess(new Uri(url), streamId, token, progress, default, new(skipCacheReceive));
 var @base = await process.Deserialize(rootId).ConfigureAwait(false);
 Console.WriteLine("Deserialized");
 Console.ReadLine();
 Console.WriteLine("Executing");
 
-var process2 = factory.CreateSerializeProcess(
+var factory2 = serviceProvider.GetRequiredService<ISerializeProcessFactory>();
+var process2 = factory2.CreateSerializeProcess(
   new Uri(url),
   streamId,
   token,
